Freeze versions of some dependencies specifically for apbears
These changes should remain in a branch solely for apbears use. We'll be
moving away from this gem for future projects, but apbears has been
pointing to HEAD of master for quite while. Breaking this off into a
branch is to help make sure that any tweaks made for other repos don't
break apbears.